Add mainpage.html with layout and login forms
Started a look for the mainpage just so we can get some of the functional things going. Have a function for a fake log in - the user and password are just teacher or student and password. Just a start and we can start working on having the links between multiple pages working as well - can also move the css to another file in the future